requirements.txt
fastai>=2.7.12
pandas==2.1.4
scikit-learn
numpy
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3 